refactor: less allocations for money formatting

+++ app/src/main/java/ch/giantific/qwittig/presentation/home/purchases/addedit/PurchaseAddActivity.java
@@ -220,7 +220,7 @@ public class PurchaseAddActivity extends BaseActivity<PurchaseAddEditViewModel> 
     }
 
     @Override
-    public void onExchangeRateManuallySet(float exchangeRate) {
+    public void onExchangeRateManuallySet(double exchangeRate) {
         mViewModel.onExchangeRateManuallySet(exchangeRate);
     }
